refactor(skeleton): inline state interface
