+++ packages/components/src/components/_skeleton/component.tsx
@@ -1,7 +1,6 @@
 import type { JSX } from '@stencil/core';
 import { Component, h, Prop, State, Watch } from '@stencil/core';
-import { SkeletonFC } from './internal/functional-components/skeleton/component';
-import type { SkeletonState } from './internal/functional-components/skeleton/state';
+import { SkeletonFC, type SkeletonState } from './internal/functional-components/skeleton/component';
 import { SkeletonController, type SkeletonControllerWatchers } from './internal/functional-components/skeleton/controller';
 import type { NameProp, NamePropType } from './internal/functional-components/skeleton/schema/props/name';
 import type { ShowProp, ShowPropType } from './internal/functional-components/skeleton/schema/props/show';
@@ -35,6 +34,6 @@ export class Skeleton implements SkeletonProps, SkeletonState {
 	}
 
 	public render(): JSX.Element {
-		return <SkeletonFC nameState={this.nameState} />;
+		return <SkeletonFC nameState={this.nameState} showState={this.showState} />;
 	}
 }
