+++ packages/components/src/components/_skeleton/internal/functional-components/skeleton/component.tsx
@@ -1,6 +1,12 @@
 import type { FunctionalComponent as FC } from '@stencil/core';
 import { h } from '@stencil/core';
-import type { SkeletonState } from './state';
+import type { NamePropType } from './schema/props/name';
+import type { ShowPropType } from './schema/props/show';
+
+export interface SkeletonState {
+	nameState?: NamePropType;
+	showState?: ShowPropType;
+}
 
 export const SkeletonFC: FC<SkeletonState> = ({ nameState, showState }) => {
 	return showState && <span>{nameState}</span>;
